error messages

+++ bin/cli.js
@@ -4,36 +4,47 @@ console.log('Hello');
 var shell = require("shelljs");
 var parseArgs = require("minimist")
 
-shell.exec("echo shell.exec works");
+go();
 
-//import shell from 'shelljs';
-//import parseArgs from 'minimist';
+function go() {
 
-let args = parseArgs(process.argv.slice(2));
-const command = args._[0];
+  // Delete dist
+  shell.rm('-rf', 'dist/*')
 
-const project = 'my-demo';
+  // Get params
+  let args = parseArgs(process.argv.slice(2));
+  const command = args._[0];
+  const template = args._[1];
+  const project = args._[2];
 
-// Delete dist
-shell.rm('-rf', 'dist/*')
+  console.log('>>>>>> args ' + args._.length);
+  const projectPath = 'src/' + project;
+
+  if (!command) {
+    console.log('No command was entered.  Supported commnads are g (generate) only.');
+    return;
+  }
+  if (!template) {
+    console.log('No template was entered.  You must chose a template to base your project from.  eg "project"');
+    return;
+  }
+  if (!project) {
+    console.log('No project name was entered.  You must chose a new name for your project.');
+    return;
+  }
+  
+  // Copy template project to new project
+  shell.rm('-rf', projectPath)
+  shell.mkdir(projectPath);
+  shell.cp('-rf', '_templates/' + template + '/*', projectPath);
+
+  // rename project.json
+  shell.mv(projectPath + '/project.json', projectPath + '/' + project + '.json');
+
+  // Generate project
+  shell.exec("npx @11ty/eleventy --serve");
+
+}
 
 
-const projectPath = 'src/' + project;
-shell.rm('-rf', projectPath)
-shell.mkdir(projectPath);
-shell.cp('-rf', '_templates/project/*', projectPath);
 
-// rename project.json
-shell.mv(projectPath + '/project.json', projectPath + '/' + project + '.json');
-
-shell.cd(projectPath);
-// shell.ls('*.*').forEach(function (file) {
-//   shell.sed('-i', 'PAGE_TITLE', 'New Title for ' + project, file);
-//   shell.sed('-i', 'SITE_NAME', project, file);
-// });
-shell.cd('../..');
-
-// Generate project
-shell.exec("npx @11ty/eleventy --serve");
-
-console.log('>>>>>> command ' + command);
